requirements.txt
pandas
numpy
matplotlib
seaborn
plotly
scipy
scikit-learn
statsmodels
cdsapi
python-dotenv
requests
